+++ client/src/API/UserLayout.js
@@ -1,12 +1,15 @@
-import { Link, Route, Switch } from "react-router-dom";
 import React from "react";
+
+//import react router
+import { Link, Route, Switch } from "react-router-dom";
+
 
 // Page Import
 import UserDetail from "./UserDetailPage";
 
 function UserLayout({ userinfos }) {
   const src = "/api/userinfo/img/";
-
+//if youre  creating a classname please camelcase, so instead of linkimg do linkImg its just easier to read
   const userElement = userinfos.map(userinfo => (
     <div className={"likeProfile"}>
       <Link className={"likeLink"} to={`userinfo/Matches/${userinfo._id}`}>
