+++ package.json
@@ -17,7 +17,6 @@
   "bugs": {
     "url": "https://github.com/nickolipengelly/keyvaluepairs/issues"
   },
-  "proxy": "http://localhost:8700",
   "homepage": "https://github.com/nickolipengelly/keyvaluepairs#readme",
   "dependencies": {
     "body-parser": "^1.18.3",
